Added most upto date pyk kernrels from lassen As cuda only works on lassen I have to devlope there for these

+++ mcdc_tnt/pyk_kernels/all/cleanup.py
@@ -10,17 +10,17 @@ import numpy as np
 @pk.workload
 class BringOutYourDead:
     def __init__ (self,p_pos_x, p_pos_y, p_pos_z, p_mesh_cell, p_dir_y, p_dir_z, p_dir_x, p_speed, p_time, p_alive, num_part, clever_out):
-        self.p_pos_x: pk.View1D[pk.double] = p_pos_x
-        self.p_pos_y: pk.View1D[pk.double] = p_pos_y
-        self.p_pos_z: pk.View1D[pk.double] = p_pos_z
+        self.p_pos_x: pk.View1D[pk.float] = p_pos_x
+        self.p_pos_y: pk.View1D[pk.float] = p_pos_y
+        self.p_pos_z: pk.View1D[pk.float] = p_pos_z
         
-        self.p_dir_x: pk.View1D[pk.double] = p_dir_x
-        self.p_dir_y: pk.View1D[pk.double] = p_dir_y
-        self.p_dir_z: pk.View1D[pk.double] = p_dir_z
+        self.p_dir_x: pk.View1D[pk.float] = p_dir_x
+        self.p_dir_y: pk.View1D[pk.float] = p_dir_y
+        self.p_dir_z: pk.View1D[pk.float] = p_dir_z
         
         self.p_mesh_cell: pk.View1D[int] = p_mesh_cell
-        self.p_speed: pk.View1D[pk.double] = p_speed
-        self.p_time: pk.View1D[pk.double] = p_time
+        self.p_speed: pk.View1D[pk.float] = p_speed
+        self.p_time: pk.View1D[pk.float] = p_time
         self.p_alive: pk.View1D[int] = p_alive
         
         self.num_part: int = num_part
